refactor: move user table handling
