+++ includes/class-users.php
@@ -1,16 +1,14 @@
 <?php
 /**
- * Newspack Hub Users Admin page
+ * Newspack Users Admin page
  *
  * @package Newspack
  */
 
-namespace Newspack_Network\Hub\Admin;
-
-use Newspack_Network\Hub\Stores\Event_Log as Event_Log_Store;
+namespace Newspack_Network;
 
 /**
- * Class to handle the Users admin page by adding an additional column to the Users table
+ * Class to handle the Users admin page
  */
 class Users {
 
@@ -20,6 +18,7 @@ class Users {
 	public static function init() {
 		add_filter( 'manage_users_columns', [ __CLASS__, 'manage_users_columns' ] );
 		add_filter( 'manage_users_custom_column', [ __CLASS__, 'manage_users_custom_column' ], 10, 3 );
+		add_filter( 'users_list_table_query_args', [ __CLASS__, 'users_list_table_query_args' ] );
 	}
 
 	/**
@@ -29,7 +28,9 @@ class Users {
 	 * @return array
 	 */
 	public static function manage_users_columns( $columns ) {
-		$columns['newspack_network_activity'] = __( 'Newspack Network Activity', 'newspack-network' );
+		if ( Site_Role::is_hub() ) {
+			$columns['newspack_network_activity'] = __( 'Newspack Network Activity', 'newspack-network' );
+		}
 		return $columns;
 	}
 
@@ -42,13 +43,13 @@ class Users {
 	 * @return string
 	 */
 	public static function manage_users_custom_column( $value, $column_name, $user_id ) {
-		if ( 'newspack_network_activity' === $column_name ) {
+		if ( 'newspack_network_activity' === $column_name && Site_Role::is_hub() ) {
 			$user = get_user_by( 'id', $user_id );
 			if ( ! $user ) {
 				return $value;
 			}
 
-			$last_activity = Event_Log_Store::get( [ 'email' => $user->user_email ], 1 );
+			$last_activity = \Newspack_Network\Hub\Stores\Event_Log::get( [ 'email' => $user->user_email ], 1 );
 
 			if ( empty( $last_activity ) ) {
 				return '-';
@@ -59,7 +60,7 @@ class Users {
 			$summary       = $last_activity->get_summary();
 			$event_log_url = add_query_arg(
 				[
-					'page'  => Event_Log::PAGE_SLUG,
+					'page'  => \Newspack_Network\Hub\Admin\Event_Log::PAGE_SLUG,
 					'email' => $user->user_email,
 				],
 				admin_url( 'admin.php' )
@@ -74,5 +75,20 @@ class Users {
 
 		}
 		return $value;
+	}
+
+	/**
+	 * Handle the filtering of users by multiple roles.
+	 * Unfortunatelly, `get_views` and `get_views_links` are not filterable, so "All" will
+	 * be displayed as the active filter.
+	 *
+	 * @param array $args The current query args.
+	 */
+	public static function users_list_table_query_args( $args ) {
+		if ( isset( $_REQUEST['role__in'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+			$args['role__in'] = explode( ',', sanitize_text_field( $_REQUEST['role__in'] ) ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+			unset( $args['role'] );
+		}
+		return $args;
 	}
 }
