+++ includes/hub/class-admin.php
@@ -20,7 +20,6 @@ class Admin {
 		Admin\Subscriptions::init();
 		Admin\Orders::init();
 		Admin\Membership_Plans::init();
-		Admin\Users::init();
 		Admin\Nodes_List::init();
 		Distributor_Settings::init();
 	}
